move files to DataType
